+++ tests/test_core.py
@@ -6,7 +6,7 @@ from numpy import nan
 from pandas import DataFrame, Index, MultiIndex
 from pandas.testing import assert_frame_equal, assert_index_equal
 
-from pandas_indexing.core import assignlevel, dropna
+from pandas_indexing.core import assignlevel, dropnalevel
 
 
 def test_assignlevel_index(sidx: Index, midx: MultiIndex):
@@ -87,16 +87,18 @@ def test_assignlevel_dataframe(mdf: DataFrame):
     )
 
 
-def test_dropna(mdf):
+def test_dropnalevel(mdf):
     midx = MultiIndex.from_tuples(
         [("foo", nan, nan), ("foo", nan, 2), ("bar", 3, 3)],
         names=["str", "num", "num2"],
     )
     mdf = mdf.set_axis(midx)
 
-    assert_index_equal(dropna(midx), midx[[2]])
-    assert_frame_equal(dropna(mdf), mdf.iloc[[2]])
+    assert_index_equal(dropnalevel(midx), midx[[2]])
+    assert_frame_equal(dropnalevel(mdf), mdf.iloc[[2]])
 
-    assert_index_equal(dropna(midx, subset=["num", "num2"], how="all"), midx[[1, 2]])
-    assert_index_equal(dropna(midx, subset=["str", "num"]), midx[[2]])
-    assert_index_equal(dropna(midx, subset="num2"), midx[[1, 2]])
+    assert_index_equal(
+        dropnalevel(midx, subset=["num", "num2"], how="all"), midx[[1, 2]]
+    )
+    assert_index_equal(dropnalevel(midx, subset=["str", "num"]), midx[[2]])
+    assert_index_equal(dropnalevel(midx, subset="num2"), midx[[1, 2]])